refactor(agent): own the engine contract in one leaf, inline vocab's conversation wrapper
The client engine and the workspace loop each carried their own hand-copied
"structural twin" of the wire contract (AgentEngine, EngineChunk, ModelMessage,
and friends) because neither package may depend on the other's heavy graph. Two
copies kept in sync by hand drift silently: a divergence surfaces as a confusing
assignability error far from its cause, with nothing recording that they were
meant to be one type.

Extract @epicenter/agent-protocol, a leaf whose only dependency is
wellcrafted/json, as the single home for that contract. Both sides import it;
the workspace agent re-exports it (public surface unchanged) and the client adds
only EngineFetch, the one genuinely client-side transport type. The wire shape
can no longer drift between producer and consumer.

Also inline vocab's one-caller createConversation wrapper into ConversationView.
It was a three-line partial application whose main cost was a name collision:
createConversation meant the workspace loop, this binding, and an unrelated
conversations-table row creator all at once.
