+++ packages/agent-protocol/package.json
@@ -1,0 +1,24 @@
+{
+	"name": "@epicenter/agent-protocol",
+	"version": "0.1.0",
+	"license": "AGPL-3.0-or-later",
+	"type": "module",
+	"description": "The wire/loop contract the client agent loop and an inference engine agree on: prompt messages, streamed chunks, and the engine function shape (ADR-0049/0050).",
+	"exports": {
+		".": "./src/index.ts"
+	},
+	"files": [
+		"src",
+		"!src/**/*.test.ts"
+	],
+	"dependencies": {
+		"wellcrafted": "catalog:"
+	},
+	"devDependencies": {
+		"@types/bun": "catalog:",
+		"typescript": "catalog:"
+	},
+	"scripts": {
+		"typecheck": "tsc --noEmit"
+	}
+}
